Fixed some errors in the controllers but have questions on the data access layer and the like

+++ src/main/java/com/cardealer/Repository/CustomerRepository.java
@@ -1,10 +1,12 @@
 package com.cardealer.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 import com.cardealer.Model.Customer;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository{
+public interface CustomerRepository {
  
 	int save(Customer customer);
 
@@ -14,6 +16,6 @@ public interface CustomerRepository{
 
     List<Customer> findAll();
     
-    Customer findById(Long id);
+    Customer findById(int id);
 
 }
